Restructure handler into packages
