+++ services/m3coordinator/handler/namespace/add.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package handler
+package namespace
 
 import (
 	"context"
@@ -30,6 +30,7 @@ import (
 	"time"
 
 	"github.com/m3db/m3coordinator/generated/proto/admin"
+	"github.com/m3db/m3coordinator/services/m3coordinator/handler"
 	"github.com/m3db/m3coordinator/util/logging"
 
 	m3clusterClient "github.com/m3db/m3cluster/client"
@@ -42,8 +43,8 @@ import (
 )
 
 const (
-	// NamespaceAddURL is the url for the placement add handler (with the POST method).
-	NamespaceAddURL = "/namespace/add"
+	// AddURL is the url for the placement add handler (with the POST method).
+	AddURL = "/namespace/add"
 
 	defaultBlockDataExpiryPeriodStr = "5m"
 )
@@ -52,31 +53,31 @@ var (
 	errMissingNamespaceName = errors.New("must specify namespace name")
 )
 
-// namespaceAddHandler represents a handler for placement add endpoint.
-type namespaceAddHandler AdminHandler
+// addHandler represents a handler for placement add endpoint.
+type addHandler Handler
 
-// NewNamespaceAddHandler returns a new instance of handler.
-func NewNamespaceAddHandler(clusterClient m3clusterClient.Client) http.Handler {
-	return &namespaceAddHandler{
+// NewAddHandler returns a new instance of handler.
+func NewAddHandler(clusterClient m3clusterClient.Client) http.Handler {
+	return &addHandler{
 		clusterClient: clusterClient,
 	}
 }
 
-func (h *namespaceAddHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+func (h *addHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	logger := logging.WithContext(ctx)
 
 	req, rErr := h.parseRequest(r)
 	if rErr != nil {
 		logger.Error("unable to parse request", zap.Any("error", rErr))
-		Error(w, rErr.Error(), rErr.Code())
+		handler.Error(w, rErr.Error(), rErr.Code())
 		return
 	}
 
-	nsRegistry, err := h.namespaceAdd(ctx, req)
+	nsRegistry, err := h.add(ctx, req)
 	if err != nil {
 		logger.Error("unable to get namespace", zap.Any("error", err))
-		Error(w, err, http.StatusInternalServerError)
+		handler.Error(w, err, http.StatusInternalServerError)
 		return
 	}
 
@@ -84,32 +85,32 @@ func (h *namespaceAddHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) 
 		Registry: &nsRegistry,
 	}
 
-	WriteProtoMsgJSONResponse(w, resp, logger)
+	handler.WriteProtoMsgJSONResponse(w, resp, logger)
 }
 
-func (h *namespaceAddHandler) parseRequest(r *http.Request) (*admin.NamespaceAddRequest, *ParseError) {
+func (h *addHandler) parseRequest(r *http.Request) (*admin.NamespaceAddRequest, *handler.ParseError) {
 	body, err := ioutil.ReadAll(r.Body)
 	if err != nil {
-		return nil, NewParseError(err, http.StatusBadRequest)
+		return nil, handler.NewParseError(err, http.StatusBadRequest)
 	}
 	defer r.Body.Close()
 
 	addReq := new(admin.NamespaceAddRequest)
 	if err := json.Unmarshal(body, addReq); err != nil {
-		return nil, NewParseError(err, http.StatusBadRequest)
+		return nil, handler.NewParseError(err, http.StatusBadRequest)
 	}
 
 	return addReq, nil
 }
 
-func (h *namespaceAddHandler) namespaceAdd(ctx context.Context, r *admin.NamespaceAddRequest) (nsproto.Registry, error) {
+func (h *addHandler) add(ctx context.Context, r *admin.NamespaceAddRequest) (nsproto.Registry, error) {
 	var emptyReg = nsproto.Registry{}
 	kv, err := h.clusterClient.KV()
 	if err != nil {
 		return emptyReg, err
 	}
 
-	currentMetadata, err := currentNamespaceMetadata(kv)
+	currentMetadata, err := Metadata(kv)
 	if err != nil {
 		return emptyReg, err
 	}
