+++ services/m3coordinator/handler/placement/get.go
@@ -1,0 +1,92 @@
+// Copyright (c) 2018 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package placement
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/m3db/m3coordinator/generated/proto/admin"
+	"github.com/m3db/m3coordinator/services/m3coordinator/config"
+	"github.com/m3db/m3coordinator/services/m3coordinator/handler"
+	"github.com/m3db/m3coordinator/util/logging"
+
+	m3clusterClient "github.com/m3db/m3cluster/client"
+	"github.com/m3db/m3cluster/placement"
+
+	"go.uber.org/zap"
+)
+
+const (
+	// GetURL is the url for the placement get handler (with the GET method).
+	GetURL = "/placement/get"
+
+	// GetHTTPMethodURL is the url for the placement get handler (with the GET method).
+	GetHTTPMethodURL = "/placement"
+)
+
+// getHandler represents a handler for placement get endpoint.
+type getHandler Handler
+
+// NewGetHandler returns a new instance of handler.
+func NewGetHandler(clusterClient m3clusterClient.Client, cfg config.Configuration) http.Handler {
+	return &getHandler{
+		clusterClient: clusterClient,
+		config:        cfg,
+	}
+}
+
+func (h *getHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	logger := logging.WithContext(ctx)
+
+	placement, version, err := h.get(ctx)
+	if err != nil {
+		// An error from `get` signifies "key not found", meaning there is
+		// no placement and as such, should not be treated as an actual error to
+		// the user.
+		w.Write([]byte("no placement found\n"))
+		return
+	}
+
+	placementProto, err := placement.Proto()
+	if err != nil {
+		logger.Error("unable to get placement protobuf", zap.Any("error", err))
+		handler.Error(w, err, http.StatusInternalServerError)
+		return
+	}
+
+	resp := &admin.PlacementGetResponse{
+		Placement: placementProto,
+		Version:   int32(version),
+	}
+
+	handler.WriteProtoMsgJSONResponse(w, resp, logger)
+}
+
+func (h *getHandler) get(ctx context.Context) (placement.Placement, int, error) {
+	ps, err := PlacementService(h.clusterClient, h.config)
+	if err != nil {
+		return nil, 0, err
+	}
+
+	return ps.Placement()
+}
