+++ services/m3coordinator/handler/placement/init.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package handler
+package placement
 
 import (
 	"context"
@@ -27,6 +27,8 @@ import (
 	"net/http"
 
 	"github.com/m3db/m3coordinator/generated/proto/admin"
+	"github.com/m3db/m3coordinator/services/m3coordinator/config"
+	"github.com/m3db/m3coordinator/services/m3coordinator/handler"
 	"github.com/m3db/m3coordinator/util/logging"
 
 	m3clusterClient "github.com/m3db/m3cluster/client"
@@ -36,41 +38,42 @@ import (
 )
 
 const (
-	// PlacementInitURL is the url for the placement init handler (with the POST method).
-	PlacementInitURL = "/placement/init"
+	// InitURL is the url for the placement init handler (with the POST method).
+	InitURL = "/placement/init"
 )
 
-// placementInitHandler represents a handler for placement init endpoint.
-type placementInitHandler AdminHandler
+// initHandler represents a handler for placement init endpoint.
+type initHandler Handler
 
-// NewPlacementInitHandler returns a new instance of handler.
-func NewPlacementInitHandler(clusterClient m3clusterClient.Client) http.Handler {
-	return &placementInitHandler{
+// NewInitHandler returns a new instance of handler.
+func NewInitHandler(clusterClient m3clusterClient.Client, cfg config.Configuration) http.Handler {
+	return &initHandler{
 		clusterClient: clusterClient,
+		config:        cfg,
 	}
 }
 
-func (h *placementInitHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+func (h *initHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	logger := logging.WithContext(ctx)
 
 	req, rErr := h.parseRequest(r)
 	if rErr != nil {
-		Error(w, rErr.Error(), rErr.Code())
+		handler.Error(w, rErr.Error(), rErr.Code())
 		return
 	}
 
-	placement, err := h.placementInit(ctx, req)
+	placement, err := h.init(ctx, req)
 	if err != nil {
 		logger.Error("unable to initialize placement", zap.Any("error", err))
-		Error(w, err, http.StatusInternalServerError)
+		handler.Error(w, err, http.StatusInternalServerError)
 		return
 	}
 
 	placementProto, err := placement.Proto()
 	if err != nil {
 		logger.Error("unable to get placement protobuf", zap.Any("error", err))
-		Error(w, err, http.StatusInternalServerError)
+		handler.Error(w, err, http.StatusInternalServerError)
 		return
 	}
 
@@ -78,25 +81,25 @@ func (h *placementInitHandler) ServeHTTP(w http.ResponseWriter, r *http.Request)
 		Placement: placementProto,
 	}
 
-	WriteProtoMsgJSONResponse(w, resp, logger)
+	handler.WriteProtoMsgJSONResponse(w, resp, logger)
 }
 
-func (h *placementInitHandler) parseRequest(r *http.Request) (*admin.PlacementInitRequest, *ParseError) {
+func (h *initHandler) parseRequest(r *http.Request) (*admin.PlacementInitRequest, *handler.ParseError) {
 	body, err := ioutil.ReadAll(r.Body)
 	if err != nil {
-		return nil, NewParseError(err, http.StatusBadRequest)
+		return nil, handler.NewParseError(err, http.StatusBadRequest)
 	}
 	defer r.Body.Close()
 
 	initReq := new(admin.PlacementInitRequest)
 	if err := json.Unmarshal(body, initReq); err != nil {
-		return nil, NewParseError(err, http.StatusBadRequest)
+		return nil, handler.NewParseError(err, http.StatusBadRequest)
 	}
 
 	return initReq, nil
 }
 
-func (h *placementInitHandler) placementInit(ctx context.Context, r *admin.PlacementInitRequest) (placement.Placement, error) {
+func (h *initHandler) init(ctx context.Context, r *admin.PlacementInitRequest) (placement.Placement, error) {
 	ps, err := PlacementService(h.clusterClient, h.config)
 	if err != nil {
 		return nil, err
