+++ services/m3coordinator/handler/placement/remove_test.go
@@ -1,4 +1,24 @@
-package handler
+// Copyright (c) 2018 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package placement
 
 import (
 	"errors"
@@ -9,6 +29,7 @@ import (
 	"testing"
 
 	"github.com/m3db/m3cluster/placement"
+	"github.com/m3db/m3coordinator/services/m3coordinator/config"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -16,7 +37,7 @@ import (
 
 func TestPlacementRemoveHandler(t *testing.T) {
 	mockClient, mockPlacementService := SetupPlacementTest(t)
-	handler := NewPlacementRemoveHandler(mockClient)
+	handler := NewRemoveHandler(mockClient, config.Configuration{})
 
 	// Test remove success
 	w := httptest.NewRecorder()
